fix: 리베이스 이후 오류 수정
