+++ app/src/main/java/com/teamwss/websoso/data/remote/service/NovelService.kt
@@ -1,4 +1,4 @@
-package com.teamwss.websoso.data.remote.api
+package com.teamwss.websoso.data.remote.service
 
 import com.teamwss.websoso.data.remote.request.PostNovelRequest
 import com.teamwss.websoso.data.remote.response.EditNovelResponse
@@ -6,31 +6,17 @@ import com.teamwss.websoso.data.remote.response.PostNovelResponse
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.PATCH
-import retrofit2.http.POST
 import retrofit2.http.Path
 
-interface PostNovelService {
-
+interface NovelService {
     @GET("/novels/{novelId}")
     suspend fun fetchPostNovelInfo(
         @Path("novelId") novelId: Long
     ): PostNovelResponse
-
-    @GET("/novels/{novelId}")
-    suspend fun fetchEditNovelInfo(
-        @Path("novelId") novelId: Long
-    ): EditNovelResponse
-
-    @POST("/user-novels/{novelId}")
-    suspend fun postPostNovelInfo(
-        @Path("novelId") novelId: Long,
-        @Body request: PostNovelRequest,
-    ): PostNovelRequest
 
     @PATCH("/user-novels/{userNovelId}")
     suspend fun editPostNovelInfo(
         @Path("userNovelId") userNovelId: Long,
         @Body request: PostNovelRequest,
     ): PostNovelRequest
-
 }
