+++ app/src/main/java/com/teamwss/websoso/data/remote/service/UserNovelService.kt
@@ -1,8 +1,13 @@
 package com.teamwss.websoso.data.remote.service
 
-import com.teamwss.websoso.data.remote.response.UserNovelsResponse
+import com.teamwss.websoso.data.remote.request.PostNovelRequest
+import com.teamwss.websoso.data.remote.response.EditNovelResponse
+import com.teamwss.websoso.data.remote.response.PostNovelResponse
+import retrofit2.http.Body
 import retrofit2.http.GET
-import retrofit2.http.Query
+import retrofit2.http.PATCH
+import retrofit2.http.POST
+import retrofit2.http.Path
 
 interface UserNovelService {
 
@@ -14,4 +19,15 @@ interface UserNovelService {
         @Query("sortType") sortType: String
     ): UserNovelsResponse
 
+
+    @GET("/novels/{novelId}")
+    suspend fun fetchEditNovelInfo(
+        @Path("novelId") novelId: Long
+    ): EditNovelResponse
+
+    @POST("/user-novels/{novelId}")
+    suspend fun postPostNovelInfo(
+        @Path("novelId") novelId: Long,
+        @Body request: PostNovelRequest,
+    ): PostNovelRequest
 }
