+++ app/src/main/res/layout/activity_post_novel.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
+
         <variable
             name="postNovelViewModel"
             type="com.teamwss.websoso.ui.postNovel.postNovelViewModel.PostNovelViewModel" />
@@ -75,7 +76,7 @@
 
                     <ImageView
                         android:id="@+id/ivPostCover"
-                        loadRoundedCoverImage="@{postNovelViewModel.editResponse.userNovelImg}"
+                        loadRoundedCoverImage="@{postNovelViewModel.editNovelInfo.image}"
                         android:layout_width="0dp"
                         android:layout_height="0dp"
                         android:layout_marginHorizontal="116dp"
